+++ _examples/beginner/basicauth/main.go
@@ -12,10 +12,9 @@ func main() {
 	app := iris.New()
 
 	authConfig := basicauth.Config{
-		Users:      map[string]string{"myusername": "mypassword", "mySecondusername": "mySecondpassword"},
-		Realm:      "Authorization Required", // defaults to "Authorization Required"
-		ContextKey: "user",                   // defaults to "user"
-		Expires:    time.Duration(30) * time.Minute,
+		Users:   map[string]string{"myusername": "mypassword", "mySecondusername": "mySecondpassword"},
+		Realm:   "Authorization Required", // defaults to "Authorization Required"
+		Expires: time.Duration(30) * time.Minute,
 	}
 
 	authentication := basicauth.New(authConfig)
@@ -23,10 +22,7 @@ func main() {
 	// to global app.Use(authentication) (or app.UseGlobal before the .Run)
 	// to routes
 	/*
-		app.Get("/mysecret", authentication, func(ctx context.Context) {
-			username := ctx.Values().GetString("user") //  the Contextkey from the authConfig
-			ctx.Writef("Hello authenticated user: %s ", username)
-		})
+		app.Get("/mysecret", authentication, h)
 	*/
 
 	app.Get("/", func(ctx context.Context) { ctx.Redirect("/admin") })
@@ -36,23 +32,22 @@ func main() {
 	needAuth := app.Party("/admin", authentication)
 	{
 		//http://localhost:8080/admin
-		needAuth.Get("/", func(ctx context.Context) {
-			username := ctx.Values().GetString("user") //  the Contextkey from the authConfig
-			ctx.Writef("Hello authenticated user: %s from: %s", username, ctx.Path())
-		})
+		needAuth.Get("/", h)
 		// http://localhost:8080/admin/profile
-		needAuth.Get("/profile", func(ctx context.Context) {
-			username := ctx.Values().GetString("user") //  the Contextkey from the authConfig
-			ctx.Writef("Hello authenticated user: %s from: % ", username, ctx.Path())
-		})
+		needAuth.Get("/profile", h)
 
 		// http://localhost:8080/admin/settings
-		needAuth.Get("/settings", func(ctx context.Context) {
-			username := authConfig.User(ctx) // shortcut for ctx.Values().GetString("user")
-			ctx.Writef("Hello authenticated user: %s from: %s", username, ctx.Path())
-		})
+		needAuth.Get("/settings", h)
 	}
 
 	// open http://localhost:8080/admin
 	app.Run(iris.Addr(":8080"))
+}
+
+func h(ctx context.Context) {
+	username, password, _ := ctx.Request().BasicAuth()
+	// third parameter it will be always true because the middleware
+	// makes sure for that, otherwise this handler will not be executed.
+
+	ctx.Writef("%s %s:%s", ctx.Path(), username, password)
 }