+++ middleware/basicauth/config.go
@@ -13,9 +13,6 @@ import (
 const (
 	// DefaultBasicAuthRealm is "Authorization Required"
 	DefaultBasicAuthRealm = "Authorization Required"
-	// DefaultBasicAuthContextKey is the "auth"
-	// this key is used to do context.Set("user", theUsernameFromBasicAuth)
-	DefaultBasicAuthContextKey = "user"
 )
 
 // DefaultExpireTime zero time
@@ -27,18 +24,16 @@ type Config struct {
 	Users map[string]string
 	// Realm http://tools.ietf.org/html/rfc2617#section-1.2. Default is "Authorization Required"
 	Realm string
-	// ContextKey the key for ctx.GetString(...). Default is 'user'
-	ContextKey string
 	// Expires expiration duration, default is 0 never expires
 	Expires time.Duration
 }
 
 // DefaultConfig returns the default configs for the BasicAuth middleware
 func DefaultConfig() Config {
-	return Config{make(map[string]string), DefaultBasicAuthRealm, DefaultBasicAuthContextKey, 0}
+	return Config{make(map[string]string), DefaultBasicAuthRealm, 0}
 }
 
-// User returns the user from context key same as 'ctx.GetString("user")' but cannot be used by the developer, this is only here in order to understand how you can get the authenticated username
-func (c Config) User(ctx context.Context) string {
-	return ctx.Values().GetString(c.ContextKey)
+// User returns the user from context key same as  ctx.Request().BasicAuth().
+func (c Config) User(ctx context.Context) (string, string, bool) {
+	return ctx.Request().BasicAuth()
 }